Add files via upload
In this study, we analyzed the impact of ad placement on the Click-Through Rates (CTR) on a popular news website. Data collected over a period of ten days recorded the CTR for the advertising campaign of the sports shoe brand “ZoomRunners,” placed in three different locations on the website: on the left, in the center (within content), and on the right.

On each day of observation, the CTR was measured for each ad placement location, creating a dataset reflecting the variations in user responses throughout the research period. In this context, CTR was calculated as the number of clicks per one hundred impressions, providing an indicator of how effectively the ads captured user attention and prompted responsive actions.
Further analysis of this data is expected to reveal whether there are significant differences in the average CTR between ad placements in these three locations. The findings of this research not only provide insights into the effectiveness of the ZoomRunners ad placement strategy on a specific news site but also have the potential to offer general guidelines applicable to other digital advertising contexts and various website layouts. With a profound understanding of user responses to ads in specific locations, marketers can optimize their strategies to achieve better outcomes in their advertising campaigns.